Switch the Todos and setTodos state to global context and apply it directly into the Todos component to override the onClick complete Todo button
