+++ src/App.js
@@ -5,6 +5,11 @@ import TodoList from "./components/TodoList";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { blueGrey } from "@mui/material/colors";
 
+import { useState } from "react";
+import { v4 as uuidv4 } from "uuid";
+
+import { TodosContext } from "./contexts/TodosContext";
+
 // Create a theme instance.
 const theme = createTheme({
   typography: {
@@ -14,7 +19,30 @@ const theme = createTheme({
   // You can add more theme customization below.
 });
 
+const initialTodos = [
+  {
+    id: uuidv4(),
+    title: "משימה 1",
+    details: "תיאור המשימה 1",
+    completed: false,
+  },
+  {
+    id: uuidv4(),
+    title: "משימה 2",
+    details: "תיאור המשימה 2",
+    completed: true,
+  },
+  {
+    id: uuidv4(),
+    title: "משימה 3",
+    details: "תיאור המשימה 3",
+    completed: false,
+  },
+];
+
 function App() {
+  const [todos, setTodos] = useState(initialTodos);
+
   return (
     <ThemeProvider theme={theme}>
       <div
@@ -28,7 +56,9 @@ function App() {
           direction: "rtl",
         }}
       >
-        <TodoList />
+        <TodosContext.Provider value={{ todos, setTodos }}>
+          <TodoList />
+        </TodosContext.Provider>
       </div>
     </ThemeProvider>
   );
