+++ src/contexts/TodosContext.js
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+export const TodosContext = createContext({
+  todos: [],
+  setTodos: () => {},
+});
